src/holopy/quantum/state_preparation.py
"""
Quantum state preparation with holographic verification.
"""
from typing import List, Tuple, Dict, Optional
import numpy as np
from scipy.linalg import sqrtm
import logging
from dataclasses import dataclass
from .error_mitigation import HolographicMitigation
from ..config.constants import (
    INFORMATION_GENERATION_RATE,
    PLANCK_CONSTANT
)

logger = logging.getLogger(__name__)

@dataclass
class PreparationMetrics:
    """Metrics for quantum state preparation."""
    preparation_fidelity: float
    state_purity: float
    entanglement_entropy: float
    verification_confidence: float
    preparation_time: float

class HolographicPreparation:
    """Implements quantum state preparation with holographic verification."""
    
    def __init__(
        self,
        n_qubits: int,
        error_mitigation: Optional[HolographicMitigation] = None,
        verification_threshold: float = 0.95
    ):
        """
        Initialize state preparation system.
        
        Args:
            n_qubits: Number of qubits
            error_mitigation: Optional error mitigation system
            verification_threshold: Threshold for state verification
        """
        self.n_qubits = n_qubits
        self.error_mitigation = error_mitigation
        self.verification_threshold = verification_threshold
        
        # Initialize preparation parameters
        self._initialize_preparation()
        
        logger.info(f"Initialized HolographicPreparation for {n_qubits} qubits")
    
    def _initialize_preparation(self) -> None:
        """Initialize state preparation parameters."""
        try:
            # Initialize standard basis states
            self.basis_states = self._generate_basis_states()
            
            # Initialize entangling operations
            self._initialize_entangling_operations()
            
            logger.debug("Initialized preparation parameters")
            
        except Exception as e:
            logger.error(f"Preparation initialization failed: {str(e)}")
            raise
    
    def prepare_state(
        self,
        target_state: np.ndarray,
        max_attempts: int = 3
    ) -> Tuple[np.ndarray, PreparationMetrics]:
        """
        Prepare quantum state with verification.
        
        Args:
            target_state: Target quantum state
            max_attempts: Maximum preparation attempts
            
        Returns:
            Tuple of (prepared_state, preparation_metrics)
        """
        try:
            best_fidelity = 0.0
            best_state = None
            best_metrics = None
            
            for attempt in range(max_attempts):
                # Prepare state
                prepared_state = self._prepare_single_attempt(target_state)
                
                # Apply error mitigation if available
                if self.error_mitigation:
                    result = self.error_mitigation.mitigate_errors(
                        prepared_state,
                        target_state
                    )
                    prepared_state = result.mitigated_state
                
                # Verify preparation
                metrics = self._verify_preparation(
                    prepared_state,
                    target_state
                )
                
                if metrics.preparation_fidelity > best_fidelity:
                    best_fidelity = metrics.preparation_fidelity
                    best_state = prepared_state
                    best_metrics = metrics
                
                if best_fidelity >= self.verification_threshold:
                    break
                
                logger.debug(
                    f"Preparation attempt {attempt + 1} "
                    f"fidelity: {metrics.preparation_fidelity:.4f}"
                )
            
            if best_state is None:
                raise ValueError("State preparation failed all attempts")
            
            return best_state, best_metrics
            
        except Exception as e:
            logger.error(f"State preparation failed: {str(e)}")
            raise
    
    def prepare_entangled_state(
        self,
        entanglement_pattern: str
    ) -> Tuple[np.ndarray, PreparationMetrics]:
        """
        Prepare specific entangled state.
        
        Args:
            entanglement_pattern: Pattern specifying entanglement
            
        Returns:
            Tuple of (entangled_state, preparation_metrics)
        """
        try:
            # Generate target entangled state
            target_state = self._generate_entangled_state(entanglement_pattern)
            
            # Prepare state
            return self.prepare_state(target_state)
            
        except Exception as e:
            logger.error(f"Entangled state preparation failed: {str(e)}")
            raise
    
    def _prepare_single_attempt(
        self,
        target_state: np.ndarray
    ) -> np.ndarray:
        """Perform single preparation attempt."""
        try:
            # Start from |0⟩ state
            state = np.zeros(2**self.n_qubits, dtype=complex)
            state[0] = 1.0
            
            # Apply sequence of operations
            operations = self._decompose_target_state(target_state)
            
            for op in operations:
                state = op @ state
                
                # Apply holographic constraints
                state = self._apply_holographic_constraints(state)
            
            return state
            
        except Exception as e:
            logger.error(f"Single preparation attempt failed: {str(e)}")
            raise
    
    def _verify_preparation(
        self,
        prepared_state: np.ndarray,
        target_state: np.ndarray
    ) -> PreparationMetrics:
        """Verify prepared state quality."""
        try:
            # Calculate fidelity
            fidelity = np.abs(np.vdot(prepared_state, target_state))**2
            
            # Calculate purity
            density = np.outer(prepared_state, np.conj(prepared_state))
            purity = np.real(np.trace(density @ density))
            
            # Calculate entanglement entropy
            entropy = self._calculate_entanglement_entropy(prepared_state)
            
            # Calculate verification confidence
            confidence = self._calculate_verification_confidence(
                prepared_state,
                target_state
            )
            
            return PreparationMetrics(
                preparation_fidelity=fidelity,
                state_purity=purity,
                entanglement_entropy=entropy,
                verification_confidence=confidence,
                preparation_time=0.0  # TODO: Add timing
            )
            
        except Exception as e:
            logger.error(f"State verification failed: {str(e)}")
            raise
    
    def _calculate_entanglement_entropy(
        self,
        state: np.ndarray
    ) -> float:
        """Calculate entanglement entropy of state."""
        try:
            # Reshape for bipartite split
            n = len(state)
            mid = n // 2
            rho = np.outer(state, np.conj(state))
            rho_reshaped = rho.reshape(mid, n//mid, mid, n//mid)
            
            # Calculate reduced density matrix
            rho_a = np.trace(rho_reshaped, axis1=1, axis2=3)
            
            # Calculate von Neumann entropy
            eigenvals = np.linalg.eigvalsh(rho_a)
            eigenvals = eigenvals[eigenvals > 1e-10]
            entropy = -np.sum(eigenvals * np.log2(eigenvals))
            
            return entropy
            
        except Exception as e:
            logger.error(f"Entropy calculation failed: {str(e)}")
            raise
    
    def _calculate_verification_confidence(
        self,
        prepared_state: np.ndarray,
        target_state: np.ndarray
    ) -> float:
        """Calculate confidence in state verification."""
        try:
            # Calculate statistical distance
            prepared_probs = np.abs(prepared_state)**2
            target_probs = np.abs(target_state)**2
            
            statistical_distance = np.sum(
                np.abs(prepared_probs - target_probs)
            ) / 2
            
            # Calculate confidence based on distance
            confidence = np.exp(-statistical_distance / 
                              INFORMATION_GENERATION_RATE)
            
            return confidence
            
        except Exception as e:
            logger.error(f"Confidence calculation failed: {str(e)}")
            raise 